+++ api-owner/src/main/java/com/wellmeet/owner/OwnerController.java
@@ -1,0 +1,11 @@
+package com.wellmeet.owner;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class OwnerController {
+
+    private final OwnerService ownerService;
+}
