+++ api-owner/src/main/java/com/wellmeet/owner/OwnerService.java
@@ -1,4 +1,4 @@
-package com.wellmeet.ownermanage;
+package com.wellmeet.owner;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
