--- api-owner/src/main/java/com/wellmeet/ownermanage/owner/OwnerService.java
@@ -1,9 +1,0 @@
-package com.wellmeet.ownermanage.owner;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class OwnerService {
-}
